Implemented SharedGateway\AbstractGateway to avoid code duplication
