+++ src/lib/Persistence/Legacy/SharedGateway/DatabasePlatform/SqliteGateway.php
@@ -9,9 +9,8 @@ declare(strict_types=1);
 namespace Ibexa\Core\Persistence\Legacy\SharedGateway\DatabasePlatform;
 
 use Ibexa\Core\Base\Exceptions\DatabaseException;
-use Ibexa\Core\Persistence\Legacy\SharedGateway\Gateway;
 
-final class SqliteGateway implements Gateway
+final class SqliteGateway extends AbstractGateway
 {
     /**
      * Error code 7 for a fatal error - taken from an existing driver implementation.
@@ -21,7 +20,7 @@ final class SqliteGateway implements Gateway
     private const DB_INT_MAX = 2147483647;
 
     /** @var array<string, int> */
-    private $lastInsertedIds = [];
+    private array $lastInsertedIds = [];
 
     public function getColumnNextIntegerValue(
         string $tableName,
@@ -48,28 +47,6 @@ final class SqliteGateway implements Gateway
         }
 
         return $this->lastInsertedIds[$sequenceName];
-    }
-
-    /**
-     * Return a language sub select query for setName.
-     *
-     * The query generates the proper language mask at the runtime of the INSERT/UPDATE query
-     * generated by setName.
-     *
-     * @see setName
-     */
-    public function getSetNameLanguageMaskSubQuery(): string
-    {
-        return <<<SQL
-            (SELECT
-                CASE
-                    WHEN (initial_language_id = :language_id AND (language_mask & :language_id) <> 0 )
-                    THEN (:language_id | 1)
-                    ELSE :language_id
-                END
-                FROM ezcontentobject
-                WHERE id = :content_id)
-            SQL;
     }
 }
