+++ src/lib/Resources/settings/storage_engines/legacy/shared_gateway.yaml
@@ -6,6 +6,8 @@ services:
         bind:
             $connection: '@ibexa.persistence.connection'
 
+    Ibexa\Core\Persistence\Legacy\SharedGateway\DatabasePlatform\AbstractGateway: ~
+
     Ibexa\Core\Persistence\Legacy\SharedGateway\DatabasePlatform\FallbackGateway: ~
 
     Ibexa\Core\Persistence\Legacy\SharedGateway\DatabasePlatform\SqliteGateway:
